refactor: whisper api 요청 코드 분리
