--- src/test/java/com/example/memic/service/AudioTranscriptionServiceTest.java
@@ -1,15 +1,0 @@
-package com.example.memic.service;
-
-class AudioTranscriptionServiceTest {
-
-//    @Test
-//    void mp3_자막추출() {
-//        final var service = new TranscriptionService();
-//        ResponseEntity<String> response = service.transcribe("/Users/jung-yunho/youtube-mp3/lpcpsCY4Mco.mp3");
-//        SoftAssertions.assertSoftly(
-//                softly -> {
-//                    softly.assertThat(response.getStatusCode().value()).isEqualTo(HttpStatus.OK.value());
-//                }
-//        );
-//    }
-}
